Output on stderr to not get muted when the process is killed on timeout

+++ Benchmarks/runner.py
@@ -143,11 +143,13 @@ for filename in args.filenames:
             print("Wait up to {} seconds".format(args.timeout))
             sys.stdout.flush()
             p.join(args.timeout)
-            p.terminate()
+            if p.is_alive():
+                print("Timeout!")
+                p.terminate()
             try:
                 ans = q.get(block=False)
             except queue.Empty:
-                ans = 'RSF'
+                ans = 'timeout'
         elif args.x == 'civl':
             ans = runner_civl.civlrun(cmd, args.timeout, filename, binary, test_count)
 #        elif args.x == 'simgrid':
